requirements.txt
numpy
scipy
matplotlib
pillow
tqdm
colorlog
pyyaml